Re-organized the ast-toolkit-span's internals
Separated out the crate into modules, and separated some `Spannable` features into separate traits.
