+++ lib/ast-toolkit-snack/src/sequence.rs
@@ -4,7 +4,7 @@
 //  Created:
 //    05 Apr 2024, 13:35:22
 //  Last edited:
-//    03 May 2024, 17:09:16
+//    06 May 2024, 11:55:05
 //  Auto updated?
 //    Yes
 //
@@ -24,19 +24,33 @@ use crate::{Combinator, Expects, ExpectsFormatter, Result};
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::utf8::complete::tag;
+    use crate::span::MatchBytes;
+    use crate::utf8::complete::{tag, Tag};
 
     #[test]
     fn multiple_lifetimes() {
-        // We have two tags, one with infinite lifetime and one with less
-        let shorter = String::from("shorter");
-        let tag1 = tag("STATIC");
-        let tag2 = tag(shorter.as_str());
+        // fn two_tags<'t, F: Clone, S: Clone + MatchBytes>(t: &'t str) -> Tuple<F, S, (Tag<'static, F, S>, Tag<'t, F, S>)> {
+        //     tuple((tag("STATIC"), tag(t)))
+        // }
+        // fn pre_tags<'t, F: Clone, S: Clone + MatchBytes>(t: &'t str) -> Preceded<F, S, Tag<'static, F, S>, Tag<'t, F, S>> {
+        //     preceded(tag("STATIC"), tag(t))
+        // }
+        // fn pre_tags_rev<'t, F: Clone, S: Clone + MatchBytes>(t: &'t str) -> Preceded<F, S, Tag<'t, F, S>, Tag<'static, F, S>> {
+        //     preceded(tag(t), tag("STATIC"))
+        // }
 
-        // Get a pair of things
-        let mut comb = terminated(tag1, tag2);
-        let span = Span::new("<example>", "STATICshorter");
-        assert_eq!(comb.parse(span).unwrap(), (span.slice(13..), span.slice(..6)));
+        // Define various type of input
+        let span1 = Span::new("<example>", "STATICshorter");
+        let span2 = Span::new("<example>", "shorterSTATIC");
+        let t: String = String::from("shorter");
+
+        // Run the two combinators
+        // let mut comb1 = two_tags(&t);
+        // let mut comb2 = pre_tags(&t);
+        // let mut comb3 = pre_tags_rev(&t);
+        // assert_eq!(comb1.parse(span1).unwrap(), (span1.slice(13..), (span1.slice(..6), span1.slice(6..13))));
+        // assert_eq!(comb2.parse(span1).unwrap(), (span1.slice(13..), span1.slice(..6)));
+        // assert_eq!(comb3.parse(span2).unwrap(), (span2.slice(13..), span2.slice(7..13)));
     }
 }
 
@@ -78,10 +92,10 @@ mod tests {
 /// assert!(matches!(comb.parse(span3), SResult::Fail(Failure::Common(Common::Digit1 { .. }))));
 /// ```
 #[inline]
-pub const fn pair<'c, F, S, E, C1, C2>(first: C1, second: C2) -> Tuple<F, S, (C1, C2)>
+pub const fn pair<'t1, 't2, F, S, E, C1, C2>(first: C1, second: C2) -> Tuple<F, S, (C1, C2)>
 where
-    C1: Combinator<'c, F, S, Error = E>,
-    C2: Combinator<'c, F, S, Error = E>,
+    C1: Combinator<'t1, F, S, Error = E>,
+    C2: Combinator<'t2, F, S, Error = E>,
 {
     Tuple { tuple: (first, second), _f: PhantomData, _s: PhantomData }
 }
@@ -124,9 +138,9 @@ where
 /// assert!(matches!(comb.parse(span3), SResult::Fail(Failure::Common(Common::Digit1 { .. }))));
 /// ```
 #[inline]
-pub const fn tuple<'c, F, S, C>(combs: C) -> Tuple<F, S, C>
+pub const fn tuple<'t, F, S, C>(combs: C) -> Tuple<F, S, C>
 where
-    C: Combinator<'c, F, S>,
+    C: Combinator<'t, F, S>,
 {
     Tuple { tuple: combs, _f: PhantomData, _s: PhantomData }
 }
@@ -163,10 +177,10 @@ where
 /// assert!(matches!(comb.parse(span3), SResult::Fail(Failure::Common(Common::Digit1 { .. }))));
 /// ```
 #[inline]
-pub const fn terminated<'c, F, S, E, C1, C2>(first: C1, sep: C2) -> Terminated<F, S, C1, C2>
+pub const fn terminated<'t1, 't2, F, S, E, C1, C2>(first: C1, sep: C2) -> Terminated<F, S, C1, C2>
 where
-    C1: Combinator<'c, F, S, Error = E>,
-    C2: Combinator<'c, F, S, Error = E>,
+    C1: Combinator<'t1, F, S, Error = E>,
+    C2: Combinator<'t2, F, S, Error = E>,
 {
     Terminated { first, sep, _f: PhantomData, _s: PhantomData }
 }
@@ -203,10 +217,10 @@ where
 /// assert!(matches!(comb.parse(span3), SResult::Fail(Failure::Common(Common::Digit1 { .. }))));
 /// ```
 #[inline]
-pub const fn preceded<'c, F, S, E, C1, C2>(sep: C1, second: C2) -> Preceded<F, S, C1, C2>
+pub const fn preceded<'t1, 't2, F, S, E, C1, C2>(sep: C1, second: C2) -> Preceded<F, S, C1, C2>
 where
-    C1: Combinator<'c, F, S, Error = E>,
-    C2: Combinator<'c, F, S, Error = E>,
+    C1: Combinator<'t1, F, S, Error = E>,
+    C2: Combinator<'t2, F, S, Error = E>,
 {
     Preceded { sep, second, _f: PhantomData, _s: PhantomData }
 }
@@ -255,11 +269,11 @@ where
 /// ));
 /// ```
 #[inline]
-pub const fn separated_pair<'c, F, S, E, C1, C2, C3>(first: C1, sep: C2, third: C3) -> SeparatedPair<F, S, C1, C2, C3>
+pub const fn separated_pair<'t1, 't2, 't3, F, S, E, C1, C2, C3>(first: C1, sep: C2, third: C3) -> SeparatedPair<F, S, C1, C2, C3>
 where
-    C1: Combinator<'c, F, S, Error = E>,
-    C2: Combinator<'c, F, S, Error = E>,
-    C3: Combinator<'c, F, S, Error = E>,
+    C1: Combinator<'t1, F, S, Error = E>,
+    C2: Combinator<'t2, F, S, Error = E>,
+    C3: Combinator<'t3, F, S, Error = E>,
 {
     SeparatedPair { first, sep, third, _f: PhantomData, _s: PhantomData }
 }
@@ -305,11 +319,11 @@ where
 /// ));
 /// ```
 #[inline]
-pub const fn delimited<'c, F, S, E, C1, C2, C3>(sep1: C1, second: C2, sep3: C3) -> Delimited<F, S, C1, C2, C3>
+pub const fn delimited<'t1, 't2, 't3, F, S, E, C1, C2, C3>(sep1: C1, second: C2, sep3: C3) -> Delimited<F, S, C1, C2, C3>
 where
-    C1: Combinator<'c, F, S, Error = E>,
-    C2: Combinator<'c, F, S, Error = E>,
-    C3: Combinator<'c, F, S, Error = E>,
+    C1: Combinator<'t1, F, S, Error = E>,
+    C2: Combinator<'t2, F, S, Error = E>,
+    C3: Combinator<'t3, F, S, Error = E>,
 {
     Delimited { sep1, second, sep3, _f: PhantomData, _s: PhantomData }
 }
@@ -335,6 +349,30 @@ impl<'t> Display for PairExpects<'t> {
     }
 }
 impl<'t> ExpectsFormatter for PairExpects<'t> {
+    #[inline]
+    fn expects_fmt(&self, f: &mut Formatter, indent: usize) -> FResult {
+        self.first_fmt.expects_fmt(f, indent)?;
+        write!(f, ", then ")?;
+        self.second_fmt.expects_fmt(f, indent)
+    }
+}
+
+/// ExpectsFormatter for two consequtive other formatters.
+#[derive(Debug)]
+pub struct PairExpectsBeta<E1, E2> {
+    /// The first thing we expect.
+    first_fmt:  E1,
+    /// The second thing we expect.
+    second_fmt: E2,
+}
+impl<E1: ExpectsFormatter, E2: ExpectsFormatter> Display for PairExpectsBeta<E1, E2> {
+    #[inline]
+    fn fmt(&self, f: &mut Formatter<'_>) -> FResult {
+        write!(f, "Expected ")?;
+        self.expects_fmt(f, 0)
+    }
+}
+impl<E1: ExpectsFormatter, E2: ExpectsFormatter> ExpectsFormatter for PairExpectsBeta<E1, E2> {
     #[inline]
     fn expects_fmt(&self, f: &mut Formatter, indent: usize) -> FResult {
         self.first_fmt.expects_fmt(f, indent)?;
@@ -391,12 +429,12 @@ impl<'t, F, S, T: Expects<'t>> Expects<'t> for Tuple<F, S, T> {
     #[inline]
     fn expects(&self) -> Self::Formatter { self.tuple.expects() }
 }
-impl<'c, F, S, T: Combinator<'c, F, S>> Combinator<'c, F, S> for Tuple<F, S, T> {
+impl<'t, F, S, T: Combinator<'t, F, S>> Combinator<'t, F, S> for Tuple<F, S, T> {
     type Output = T::Output;
     type Error = T::Error;
 
     #[inline]
-    fn parse(&mut self, input: Span<F, S>) -> Result<'c, Self::Output, F, S, Self::Error> { self.tuple.parse(input) }
+    fn parse(&mut self, input: Span<F, S>) -> Result<'t, Self::Output, F, S, Self::Error> { self.tuple.parse(input) }
 }
 
 /// Combinator returned by [`terminated()`].
@@ -416,16 +454,16 @@ impl<'t, F, S, C1: Expects<'t>, C2: Expects<'t>> Expects<'t> for Terminated<F, S
     #[inline]
     fn expects(&self) -> Self::Formatter { PairExpects { first_fmt: Box::new(self.first.expects()), second_fmt: Box::new(self.sep.expects()) } }
 }
-impl<'c, F, S, E, C1, C2> Combinator<'c, F, S> for Terminated<F, S, C1, C2>
+impl<'t, F, S, E, C1, C2> Combinator<'t, F, S> for Terminated<F, S, C1, C2>
 where
-    C1: Combinator<'c, F, S, Error = E>,
-    C2: Combinator<'c, F, S, Error = E>,
+    C1: Combinator<'t, F, S, Error = E>,
+    C2: Combinator<'t, F, S, Error = E>,
 {
     type Output = C1::Output;
     type Error = E;
 
     #[inline]
-    fn parse(&mut self, input: Span<F, S>) -> Result<'c, Self::Output, F, S, Self::Error> {
+    fn parse(&mut self, input: Span<F, S>) -> Result<'t, Self::Output, F, S, Self::Error> {
         // Parse the first first
         let (rem, res): (Span<F, S>, C1::Output) = match self.first.parse(input) {
             Result::Ok(rem, res) => (rem, res),
@@ -453,22 +491,24 @@ pub struct Preceded<F, S, C1, C2> {
     /// Store the target `S`ource string type in this struct in order to be much nicer to type deduction.
     _s:     PhantomData<S>,
 }
-impl<'t, F, S, C1: Expects<'t>, C2: Expects<'t>> Expects<'t> for Preceded<F, S, C1, C2> {
-    type Formatter = PairExpects<'t>;
+impl<'t1, 't2: 't1, F, S, C1: Expects<'t1>, C2: Expects<'t2>> Expects<'t1> for Preceded<F, S, C1, C2> {
+    type Formatter = PairExpectsBeta<C1::Formatter, C2::Formatter>;
 
     #[inline]
-    fn expects(&self) -> Self::Formatter { PairExpects { first_fmt: Box::new(self.sep.expects()), second_fmt: Box::new(self.second.expects()) } }
+    fn expects(&self) -> Self::Formatter { PairExpectsBeta { first_fmt: self.sep.expects(), second_fmt: self.second.expects() } }
 }
-impl<'c, F, S, E, C1, C2> Combinator<'c, F, S> for Preceded<F, S, C1, C2>
+impl<'t1, 't2, F, S, E, C1, C2> Combinator<'t1, F, S> for Preceded<F, S, C1, C2>
 where
-    C1: Combinator<'c, F, S, Error = E>,
-    C2: Combinator<'c, F, S, Error = E>,
+    't2: 't1,
+    C1: Combinator<'t1, F, S, Error = E>,
+    C2: Combinator<'t2, F, S, Error = E>,
 {
-    type Output = C2::Output;
+    // type Output = C2::Output;
+    type Output = ();
     type Error = E;
 
     #[inline]
-    fn parse(&mut self, input: Span<F, S>) -> Result<'c, Self::Output, F, S, Self::Error> {
+    fn parse(&mut self, input: Span<F, S>) -> Result<'t1, Self::Output, F, S, Self::Error> {
         // Parse the first first
         let rem: Span<F, S> = match self.sep.parse(input) {
             Result::Ok(rem, _) => rem,
@@ -477,11 +517,12 @@ where
         };
 
         // Parse the second
-        match self.second.parse(rem) {
-            Result::Ok(rem, res) => Result::Ok(rem, res),
-            Result::Fail(fail) => Result::Fail(fail),
-            Result::Error(err) => Result::Error(err),
-        }
+        // match self.second.parse(rem) {
+        //     Result::Ok(rem, res) => Result::Ok(rem, res),
+        //     Result::Fail(fail) => Result::Fail(fail),
+        //     Result::Error(err) => Result::Error(err),
+        // }
+        Result::Ok(rem, ())
     }
 }
 
@@ -510,17 +551,17 @@ impl<'t, F, S, C1: Expects<'t>, C2: Expects<'t>, C3: Expects<'t>> Expects<'t> fo
         }
     }
 }
-impl<'c, F, S, E, C1, C2, C3> Combinator<'c, F, S> for SeparatedPair<F, S, C1, C2, C3>
+impl<'t, F, S, E, C1, C2, C3> Combinator<'t, F, S> for SeparatedPair<F, S, C1, C2, C3>
 where
-    C1: Combinator<'c, F, S, Error = E>,
-    C2: Combinator<'c, F, S, Error = E>,
-    C3: Combinator<'c, F, S, Error = E>,
+    C1: Combinator<'t, F, S, Error = E>,
+    C2: Combinator<'t, F, S, Error = E>,
+    C3: Combinator<'t, F, S, Error = E>,
 {
     type Output = (C1::Output, C3::Output);
     type Error = E;
 
     #[inline]
-    fn parse(&mut self, input: Span<F, S>) -> Result<'c, Self::Output, F, S, Self::Error> {
+    fn parse(&mut self, input: Span<F, S>) -> Result<'t, Self::Output, F, S, Self::Error> {
         // Parse the first first
         let (rem, res1): (Span<F, S>, C1::Output) = match self.first.parse(input) {
             Result::Ok(rem, res) => (rem, res),
@@ -569,17 +610,17 @@ impl<'t, F, S, C1: Expects<'t>, C2: Expects<'t>, C3: Expects<'t>> Expects<'t> fo
         }
     }
 }
-impl<'c, F, S, E, C1, C2, C3> Combinator<'c, F, S> for Delimited<F, S, C1, C2, C3>
+impl<'t, F, S, E, C1, C2, C3> Combinator<'t, F, S> for Delimited<F, S, C1, C2, C3>
 where
-    C1: Combinator<'c, F, S, Error = E>,
-    C2: Combinator<'c, F, S, Error = E>,
-    C3: Combinator<'c, F, S, Error = E>,
+    C1: Combinator<'t, F, S, Error = E>,
+    C2: Combinator<'t, F, S, Error = E>,
+    C3: Combinator<'t, F, S, Error = E>,
 {
     type Output = C2::Output;
     type Error = E;
 
     #[inline]
-    fn parse(&mut self, input: Span<F, S>) -> Result<'c, Self::Output, F, S, Self::Error> {
+    fn parse(&mut self, input: Span<F, S>) -> Result<'t, Self::Output, F, S, Self::Error> {
         // Parse the first first, discarding the result
         let rem: Span<F, S> = match self.sep1.parse(input) {
             Result::Ok(rem, _) => rem,
