--- lib/ast-toolkit-span/src/formatting/mod.rs
@@ -1,16 +1,0 @@
-//  MOD.rs
-//    by Lut99
-//
-//  Created:
-//    17 Feb 2024, 12:44:05
-//  Last edited:
-//    17 Feb 2024, 12:44:11
-//  Auto updated?
-//    Yes
-//
-//  Description:
-//!   Nested module that handles the formatting of [`Diagnostic`]s to some
-//!   writer.
-//
-
-pub mod frame;
