+++ lib/ast-toolkit-span/src/span.rs
@@ -4,7 +4,7 @@
 //  Created:
 //    15 Dec 2023, 19:05:00
 //  Last edited:
-//    03 May 2024, 14:33:42
+//    06 May 2024, 16:32:53
 //  Auto updated?
 //    Yes
 //
@@ -12,552 +12,17 @@
 //!   Implements a [`Span`], which abstracts over some input to track a particular location in it.
 //
 
-use std::borrow::Cow;
 use std::convert::Infallible;
 use std::hash::{Hash, Hasher};
 use std::ops::{Bound, RangeBounds};
 
-
-/***** HELPER MACROS *****/
-/// Converts from a [`RangeBound`] to a concrete bound.
-macro_rules! index_range_bound {
-    ($slice:expr, $range:expr) => {{
-        let slice = $slice;
-        match $range {
-            SpanRange::Closed(s, e) => &slice[s..e],
-            SpanRange::ClosedOpen(s) => &slice[s..],
-            SpanRange::OpenClosed(e) => &slice[..e],
-            SpanRange::Open => &slice[..],
-            SpanRange::Empty => &slice[0..0],
-        }
-    }};
-}
-
-
-
+use crate::eq::SpannableEq;
+use crate::hash::SpannableHash;
+use crate::range::SpanRange;
+use crate::spannable::Spannable;
 
 
 /***** AUXILLARY *****/
-/// Defines a range in some [`Span`]ned area.
-///
-/// This range can either be open- or closed on either end. 'Open' means 'everything on that side', i.e., if the left is Open, it refers to the start of the source; and if the right is Open, it refers to the end.
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub enum SpanRange {
-    /// Closed on both ends.
-    ///
-    /// Given as: inclusive start, exclusive end.
-    Closed(usize, usize),
-    /// Closed on the left, open on the right.
-    ///
-    /// Left is inclusive.
-    ClosedOpen(usize),
-    /// Open on the left, closed on the right.
-    ///
-    /// Right is exclusive.
-    OpenClosed(usize),
-    /// Fully open on both ends.
-    ///
-    /// Spans the entire source.
-    Open,
-    /// Special case of a range that doesn't span anything.
-    Empty,
-}
-impl SpanRange {
-    /// Creates a new SpanRange that is self further spanned by the given range.
-    ///
-    /// This means that the given range is assumed to be relative to this range, e.g.,
-    /// ```plain
-    /// (0..10).span(2..4) => 2..4
-    /// (5..10).span(2..4) => 7..9
-    /// (5..10).span(2..) => 7..10
-    /// ```
-    ///
-    /// # Arguments
-    /// - `span`: The other SpanRange that spans ourselves.
-    ///
-    /// # Returns
-    /// A new SpanRange that is `self` but spanned by `span`.
-    #[inline]
-    pub fn span(&self, span: &Self) -> Self {
-        // Get the simplified range of self
-        let lhs: (Option<usize>, Option<usize>) = match self {
-            Self::Closed(s, e) => (Some(*s), Some(*e)),
-            Self::ClosedOpen(s) => (Some(*s), None),
-            Self::OpenClosed(e) => (None, Some(*e)),
-            Self::Open => (None, None),
-            // If we're empty, not a lot to span
-            Self::Empty => return Self::Empty,
-        };
-
-        // Get the simplified range of other
-        let rhs: (Option<usize>, Option<usize>) = match span {
-            Self::Closed(s, e) => (Some(*s), Some(*e)),
-            Self::ClosedOpen(s) => (Some(*s), None),
-            Self::OpenClosed(e) => (None, Some(*e)),
-            Self::Open => (None, None),
-            // If the other is empty, the span is quite empty too
-            Self::Empty => return Self::Empty,
-        };
-
-        // Attempt to combine the start & end separately
-        let start: Option<usize> = match (lhs.0, rhs.0) {
-            (Some(s1), Some(s2)) => Some(s1 + s2),
-            (Some(s1), None) => Some(s1),
-            (None, Some(s2)) => Some(s2),
-            (None, None) => None,
-        };
-        let end: Option<usize> = match (lhs.0, lhs.1, rhs.1) {
-            (Some(s), Some(_), Some(e2)) => Some(s + e2),
-            (Some(_), Some(e1), None) => Some(e1),
-            (Some(s), None, Some(e2)) => Some(s + e2),
-            (Some(_), None, None) => None,
-            (None, Some(_), Some(e2)) => Some(e2),
-            (None, Some(e1), None) => Some(e1),
-            (None, None, Some(e2)) => Some(e2),
-            (None, None, None) => None,
-        };
-
-        // Now create the self
-        match (start, end) {
-            (Some(s), Some(e)) => {
-                // Double-check the range isn't empty
-                if s < e {
-                    // Double-check that both s and e are before the old one
-                    if let Some(oe) = lhs.1 {
-                        if s < oe { if e <= oe { Self::Closed(s, e) } else { Self::Closed(s, oe) } } else { Self::Empty }
-                    } else {
-                        Self::Closed(s, e)
-                    }
-                } else {
-                    Self::Empty
-                }
-            },
-            (Some(s), None) => {
-                if let Some(oe) = lhs.1 {
-                    if s < oe { Self::ClosedOpen(s) } else { Self::Empty }
-                } else {
-                    Self::ClosedOpen(s)
-                }
-            },
-            (None, Some(e)) => {
-                if let Some(oe) = lhs.1 {
-                    if e <= oe { Self::OpenClosed(e) } else { Self::OpenClosed(oe) }
-                } else {
-                    Self::OpenClosed(e)
-                }
-            },
-            (None, None) => Self::Open,
-        }
-    }
-
-    /// Creates a new SpanRange that is the merger of both self and the given one.
-    ///
-    /// # Arguments
-    /// - `other`: The other SpanRange to join with this one.
-    ///
-    /// # Returns
-    /// A new SpanRange that covers both ranges and everything in between, e.g.,
-    /// ```plain
-    /// self           other
-    /// <------>       <------------->
-    ///
-    /// result
-    /// <---------------------------->
-    /// ```
-    #[inline]
-    pub fn join(&self, other: &Self) -> Self {
-        let mut res: Self = *self;
-        res.join_mut(other);
-        res
-    }
-
-    /// Extends this SpanRange to also cover another.
-    ///
-    /// To illustrate:
-    /// ```plain
-    /// self           other
-    /// <------>       <------------->
-    ///
-    /// self.join_mut(other)
-    /// <---------------------------->
-    /// ```
-    ///
-    /// # Arguments
-    /// - `other`: The other SpanRange to join with this one.
-    #[inline]
-    pub fn join_mut(&mut self, other: &Self) {
-        // Get the simplified range of self
-        let lhs: (Option<usize>, Option<usize>) = match self {
-            Self::Closed(s, e) => (Some(*s), Some(*e)),
-            Self::ClosedOpen(s) => (Some(*s), None),
-            Self::OpenClosed(e) => (None, Some(*e)),
-            Self::Open => (None, None),
-            // If we're empty, the other determines everything
-            Self::Empty => {
-                *self = *other;
-                return;
-            },
-        };
-
-        // Get the simplified range of other
-        let rhs: (Option<usize>, Option<usize>) = match other {
-            Self::Closed(s, e) => (Some(*s), Some(*e)),
-            Self::ClosedOpen(s) => (Some(*s), None),
-            Self::OpenClosed(e) => (None, Some(*e)),
-            Self::Open => (None, None),
-            // If the other is empty, nothing needs changing
-            Self::Empty => return,
-        };
-
-        // Now compare those ones to make a proper start & end
-        let start: Option<usize> = match (lhs, rhs) {
-            // If they're both closed, we need the leftmost;
-            ((Some(s1), _), (Some(s2), _)) => Some(std::cmp::min(s1, s2)),
-            // Else, one of the lefties is open, so then so are we.
-            _ => None,
-        };
-        let end: Option<usize> = match (lhs, rhs) {
-            // If they're both closed, we need the rightmost;
-            ((_, Some(e1)), (_, Some(e2))) => Some(std::cmp::max(e1, e2)),
-            // Else, one of the righties is open, so then so are we.
-            _ => None,
-        };
-
-        // OK, finally build a new self
-        *self = match (start, end) {
-            (Some(s), Some(e)) => Self::Closed(s, e),
-            (Some(s), None) => Self::ClosedOpen(s),
-            (None, Some(e)) => Self::OpenClosed(e),
-            (None, None) => Self::Open,
-        };
-    }
-
-    /// Applies this SpanRange to a given slice.
-    ///
-    /// This slice can be anything, as long as it's a slice.
-    ///
-    /// # Arguments
-    /// - `slice`: The slice (of type [`&[T]`]) to apply this range to.
-    ///
-    /// # Returns
-    /// A new slice of type `[T]` that is the sliced counterpart.
-    #[inline]
-    pub fn apply_to<'s, T>(&self, slice: &'s [T]) -> &'s [T] { index_range_bound!(slice, *self) }
-
-    /// Applies this SpanRange to a given [`str`].
-    ///
-    /// This is the overload of [`Self::apply_to`](SpanRange::apply_to) but then to [`str`]s.
-    ///
-    /// # Arguments
-    /// - `string`: The string (of type [`&str`]) to apply this range to.
-    ///
-    /// # Returns
-    /// A new slice of type `str` that is the sliced counterpart.
-    #[inline]
-    pub fn apply_to_str<'s>(&self, string: &'s str) -> &'s str { index_range_bound!(string, *self) }
-}
-
-
-
-/// A helper trait for the [`Span`] that can be implemented for anything used as input.
-pub trait Spannable {
-    type Slice<'s>: 's
-    where
-        Self: 's;
-
-    /// Checks if this Spannable is the same (for all intends and purposes) as another Spannable of the same type.
-    ///
-    /// While it suffices to show that two [`Span`]s are the same semantically (e.g., using a simple byte-wise compare), it's also allowed to compare by pointer equality if possible for performance. Implementations for [`&[u8]`] and [`&str`] do this, for example.
-    ///
-    /// # Arguments
-    /// - `other`: Some other Spannable of type Self to check with.
-    ///
-    /// # Returns
-    /// True if these Spannables are the same, or false otherwise.
-    fn is_same(&self, other: &Self) -> bool;
-
-    /// Slices this Spannable by raw index.
-    ///
-    /// # Arguments
-    /// - `range`: The range that slices this Spannable. Can be anything implementing [`RangeBounds`].
-    ///
-    /// # Returns
-    /// A new instance of type `Self::Slice`, that is self but sliced.
-    ///
-    /// # Panics
-    /// This function panics if out-of-bounds.
-    fn slice<'s>(&'s self, range: SpanRange) -> Self::Slice<'s>;
-    /// Checks if the slices range of this Spannable is the same for the given sliced range of the same type.
-    ///
-    /// # Arguments
-    /// - `range`: The range that slices this Spannable. Can be anything implementing [`RangeBounds`].
-    /// - `other`: Some other Spannable of type Self to check with.
-    ///
-    /// # Returns
-    /// True if they are equal, or false otherwise.
-    ///
-    /// # Panics
-    /// This function panics if out-of-bounds.
-    fn slice_eq(&self, range: SpanRange, other: &Self, other_range: SpanRange) -> bool;
-    /// Computes a hash for the given range of this Spannable.
-    ///
-    /// # Arguments
-    /// - `range`: The range that slices this Spannable. Can be anything implementing [`RangeBounds`].
-    /// - `state`: Some [`Hasher`] that does the tough work.
-    ///
-    /// # Panics
-    /// This function panics if out-of-bounds.
-    fn slice_hash<H: Hasher>(&self, range: SpanRange, state: &mut H);
-
-    /// Returns the number of currently spanned "raw" items (e.g., bytes).
-    ///
-    /// # Returns
-    /// A [`usize`] with the total number of bytes or other elementary items as is stored on-disk.
-    fn byte_len(&self) -> usize;
-}
-
-// Default binary impls for [`Spannable`]
-impl<'b> Spannable for &'b [u8] {
-    type Slice<'s> = &'s [u8] where Self: 's;
-
-    #[inline]
-    fn is_same(&self, other: &Self) -> bool {
-        let ptr_eq: bool = std::ptr::eq(self, other);
-        #[cfg(debug_assertions)]
-        {
-            if !ptr_eq && self == other {
-                eprintln!(
-                    "DEBUG ASSERTION WARNING: Two byte arrays do not share the same pointer but are semantically equal. The &[u8]-implementation \
-                     for Spannable assumes comparing them by pointer equality is sufficient."
-                );
-            }
-        }
-        ptr_eq
-    }
-
-    #[inline]
-    #[track_caller]
-    fn slice<'s2>(&'s2 self, range: SpanRange) -> Self::Slice<'s2> { index_range_bound!(self, range) }
-
-    #[inline]
-    #[track_caller]
-    fn slice_eq(&self, range: SpanRange, other: &Self, other_range: SpanRange) -> bool {
-        index_range_bound!(self, range) == index_range_bound!(other, other_range)
-    }
-
-    #[inline]
-    #[track_caller]
-    fn slice_hash<H: Hasher>(&self, range: SpanRange, state: &mut H) { index_range_bound!(self, range).hash(state) }
-
-    #[inline]
-    fn byte_len(&self) -> usize { self.len() }
-}
-impl<'b, const LEN: usize> Spannable for &'b [u8; LEN] {
-    type Slice<'s> = &'s [u8] where Self: 's;
-
-    #[inline]
-    fn is_same(&self, other: &Self) -> bool {
-        let ptr_eq: bool = std::ptr::eq(self, other);
-        #[cfg(debug_assertions)]
-        {
-            if !ptr_eq && self == other {
-                eprintln!(
-                    "DEBUG ASSERTION WARNING: Two byte arrays do not share the same pointer but are semantically equal. The &[u8; \
-                     LEN]-implementation for Spannable assumes comparing them by pointer equality is sufficient."
-                );
-            }
-        }
-        ptr_eq
-    }
-
-    #[inline]
-    #[track_caller]
-    fn slice<'s2>(&'s2 self, range: SpanRange) -> Self::Slice<'s2> { index_range_bound!(self, range) }
-
-    #[inline]
-    #[track_caller]
-    fn slice_eq(&self, range: SpanRange, other: &Self, other_range: SpanRange) -> bool {
-        index_range_bound!(self, range) == index_range_bound!(other, other_range)
-    }
-
-    #[inline]
-    #[track_caller]
-    fn slice_hash<H: Hasher>(&self, range: SpanRange, state: &mut H) { index_range_bound!(self, range).hash(state) }
-
-    #[inline]
-    fn byte_len(&self) -> usize { self.len() }
-}
-impl<'b> Spannable for Cow<'b, [u8]> {
-    type Slice<'s> = Cow<'s, [u8]> where Self: 's;
-
-    #[inline]
-    fn is_same(&self, other: &Self) -> bool {
-        match (self, other) {
-            // Compare by pointer if possible
-            (Self::Borrowed(b1), Self::Borrowed(b2)) => {
-                let ptr_eq: bool = std::ptr::eq(*b1, *b2);
-                #[cfg(debug_assertions)]
-                {
-                    if !ptr_eq && self == other {
-                        eprintln!(
-                            "DEBUG ASSERTION WARNING: Two byte arrays do not share the same pointer but are semantically equal. The \
-                             Cow<u8>-implementation for Spannable assumes comparing them by pointer equality if they're both borrowed is sufficient."
-                        );
-                    }
-                }
-                ptr_eq
-            },
-            // Otherwise, fall back to equality testing
-            (o1, o2) => o1 == o2,
-        }
-    }
-
-    #[inline]
-    #[track_caller]
-    fn slice<'s2>(&'s2 self, range: SpanRange) -> Self::Slice<'s2> { Cow::Borrowed(index_range_bound!(self, range)) }
-
-    #[inline]
-    #[track_caller]
-    fn slice_eq(&self, range: SpanRange, other: &Self, other_range: SpanRange) -> bool {
-        index_range_bound!(self, range) == index_range_bound!(other, other_range)
-    }
-
-    #[inline]
-    #[track_caller]
-    fn slice_hash<H: Hasher>(&self, range: SpanRange, state: &mut H) { index_range_bound!(self, range).hash(state) }
-
-    #[inline]
-    fn byte_len(&self) -> usize { self.len() }
-}
-impl Spannable for Vec<u8> {
-    type Slice<'s> = Vec<u8> where Self: 's;
-
-    #[inline]
-    fn is_same(&self, other: &Self) -> bool { self == other }
-
-    #[inline]
-    #[track_caller]
-    fn slice<'s2>(&'s2 self, range: SpanRange) -> Self::Slice<'s2> { index_range_bound!(self, range).to_vec() }
-
-    #[inline]
-    #[track_caller]
-    fn slice_eq(&self, range: SpanRange, other: &Self, other_range: SpanRange) -> bool {
-        index_range_bound!(self, range) == index_range_bound!(other, other_range)
-    }
-
-    #[inline]
-    #[track_caller]
-    fn slice_hash<H: Hasher>(&self, range: SpanRange, state: &mut H) { index_range_bound!(self, range).hash(state) }
-
-    #[inline]
-    fn byte_len(&self) -> usize { self.len() }
-}
-
-// Default string impls for [`Spannable`]
-impl<'s> Spannable for &'s str {
-    type Slice<'s2> = &'s2 str where Self: 's2;
-
-    #[inline]
-    fn is_same(&self, other: &Self) -> bool {
-        let ptr_eq: bool = std::ptr::eq(self, other);
-        #[cfg(debug_assertions)]
-        {
-            if !ptr_eq && self == other {
-                eprintln!(
-                    "DEBUG ASSERTION WARNING: Two string slices do not share the same pointer but are semantically equal. The &str-implementation \
-                     for Spannable assumes comparing them by pointer equality is sufficient."
-                );
-            }
-        }
-        ptr_eq
-    }
-
-    #[inline]
-    #[track_caller]
-    fn slice<'s2>(&'s2 self, range: SpanRange) -> Self::Slice<'s2> { index_range_bound!(self, range) }
-
-    #[inline]
-    #[track_caller]
-    fn slice_eq(&self, range: SpanRange, other: &Self, other_range: SpanRange) -> bool {
-        index_range_bound!(self, range) == index_range_bound!(other, other_range)
-    }
-
-    #[inline]
-    #[track_caller]
-    fn slice_hash<H: Hasher>(&self, range: SpanRange, state: &mut H) { index_range_bound!(self, range).hash(state) }
-
-    #[inline]
-    fn byte_len(&self) -> usize { self.len() }
-}
-impl<'s> Spannable for Cow<'s, str> {
-    type Slice<'s2> = Cow<'s2, str> where Self: 's2;
-
-    #[inline]
-    fn is_same(&self, other: &Self) -> bool {
-        match (self, other) {
-            // Compare by pointer if possible
-            (Self::Borrowed(b1), Self::Borrowed(b2)) => {
-                let ptr_eq: bool = std::ptr::eq(*b1, *b2);
-                #[cfg(debug_assertions)]
-                {
-                    if !ptr_eq && self == other {
-                        eprintln!(
-                            "DEBUG ASSERTION WARNING: Two string slices do not share the same pointer but are semantically equal. The \
-                             Cow<str>-implementation for Spannable assumes comparing them by pointer equality if they're both borrowed is \
-                             sufficient."
-                        );
-                    }
-                }
-                ptr_eq
-            },
-            // Otherwise, fall back to equality testing
-            (o1, o2) => o1 == o2,
-        }
-    }
-
-    #[inline]
-    #[track_caller]
-    fn slice<'s2>(&'s2 self, range: SpanRange) -> Self::Slice<'s2> { Cow::Borrowed(index_range_bound!(self, range)) }
-
-    #[inline]
-    #[track_caller]
-    fn slice_eq(&self, range: SpanRange, other: &Self, other_range: SpanRange) -> bool {
-        index_range_bound!(self, range) == index_range_bound!(other, other_range)
-    }
-
-    #[inline]
-    #[track_caller]
-    fn slice_hash<H: Hasher>(&self, range: SpanRange, state: &mut H) { index_range_bound!(self, range).hash(state) }
-
-    #[inline]
-    fn byte_len(&self) -> usize { self.len() }
-}
-impl Spannable for String {
-    type Slice<'s2> = String where Self: 's2;
-
-    #[inline]
-    fn is_same(&self, other: &Self) -> bool { self == other }
-
-    #[inline]
-    #[track_caller]
-    fn slice<'s2>(&'s2 self, range: SpanRange) -> Self::Slice<'s2> { index_range_bound!(self, range).to_string() }
-
-    #[inline]
-    #[track_caller]
-    fn slice_eq(&self, range: SpanRange, other: &Self, other_range: SpanRange) -> bool {
-        index_range_bound!(self, range) == index_range_bound!(other, other_range)
-    }
-
-    #[inline]
-    #[track_caller]
-    fn slice_hash<H: Hasher>(&self, range: SpanRange, state: &mut H) { index_range_bound!(self, range).hash(state) }
-
-    #[inline]
-    fn byte_len(&self) -> usize { self.len() }
-}
-
-
-
 /// A helper trait formalizing things that return [`Span`]s.
 pub trait Spanning<F, S> {
     /// Returns some [`Span`] representing self's relation to the source text it was parsed from.
@@ -788,12 +253,12 @@ impl<F: Clone, S: Clone + Spannable> Span<F, S> {
     }
 }
 
-impl<F, S: Spannable> Eq for Span<F, S> {}
-impl<F, S: Spannable> Hash for Span<F, S> {
+impl<F, S: SpannableEq> Eq for Span<F, S> {}
+impl<F, S: SpannableHash> Hash for Span<F, S> {
     #[inline]
     fn hash<H: Hasher>(&self, state: &mut H) { self.source.slice_hash(self.range, state) }
 }
-impl<F, S: Spannable> PartialEq for Span<F, S> {
+impl<F, S: SpannableEq> PartialEq for Span<F, S> {
     #[inline]
     fn eq(&self, other: &Self) -> bool { self.source.slice_eq(self.range, &other.source, other.range) }
 }
